pyproject.toml
[tool.poetry]
name = "kasto"
version = "0.1.0"
description = "KASTO비법 위 단속 크롤링 시스템"
authors = ["mrchaos <mrchaos@naver.com>"]
license = "proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
pydantic = "^2.8.2"
undetected-chromedriver = "^3.5.5"
selenium = "^4.23.1"
webdriver-manager = "^4.0.2"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
kivy = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
